feat(d2.5-d2.8): core implementation — M1/M2/M6/M8 updates + scenario integration
Core changes:
- M1 ODD envelope manager: TMR estimator + operator state tracking
- M2 world model: COLREG-aware track aggregation + encounter classification
- M6 COLREGs reasoner: rule-based decision logic refinements
- M8 HMI transparency: SAT panel layout + real-time KPI rendering

Supporting changes:
- L3 msgs: OperatorState, ToRRequest schema v1.2.0
- Scenario: imazu-01-ho preflight gate updates for DEMO-2
- NLM: updated topic graph + citation stats
- Documentation: phase 2 alignment

Test coverage: 41 passing unit tests (M2/M5/M7/M8)

+++ src/l3_tdl_kernel/m3_mission_manager/src/mission_manager_node.cpp
@@ -539,7 +539,8 @@ void MissionManagerNode::check_current_error_severity_change(
         nlohmann::json{{"xte_nm", reading.xte_nm},
                        {"sea_current_kn", reading.sea_current_kn}});
     RCLCPP_WARN(get_logger(), "Current error HIGH: xte=%.2f nm cur=%.2f kn",
-                reading.xte_nm, reading.sea_current_kn);
+                static_cast<double>(reading.xte_nm),
+                static_cast<double>(reading.sea_current_kn));
   } else if (prev == CurrentErrorSeverity::HIGH) {
     publish_asdr_record("current_error_resolved",
         nlohmann::json{{"severity", static_cast<int>(reading.severity)}});
